+++ flask_restplus/swagger.py
@@ -188,9 +188,15 @@ class Swagger(object):
             'security': self.security_requirements(self.api.security) or None,
             'tags': tags,
             'definitions': self.serialize_definitions() or None,
-            'host': current_app.config.get('SERVER_NAME', None) or None,
+            'host': self.get_host(),
         }
         return not_none(specs)
+
+    def get_host(self):
+        hostname = current_app.config.get('SERVER_NAME', None) or None
+        if hostname and self.api.blueprint and self.api.blueprint.subdomain:
+            hostname = '.'.join((self.api.blueprint.subdomain, hostname))
+        return hostname
 
     def extract_tags(self, api):
         tags = []